pyproject.toml
[project]
name = "ynwa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "environs>=14.1.1",
    "fastapi[standard]>=0.115.12",
    "pyjwt[crypto]>=2.10.1",
    "redis>=6.0.0",
    "sqlalchemy[asyncio]>=2.0.40",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
]
